Fix(item): 상품 전체 목록 jwt 적용

+++ backend/yam/src/main/java/com/ssafy/yam/domain/item/repository/ItemRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ItemRepository extends JpaRepository<Item, Long> {
+public interface ItemRepository extends JpaRepository<Item, Integer> {
     Item findItemByItemId(int itemId);
     List<Item> findAllBy(Pageable pageable);
     List<Item> findAllBySeller_UserIdOrderByItemModifiedTime(int userId);
